reorganization of classes and modules (#31)
titles are the same size for all elements in classes and modules. They only show if there is at least one item for a category in the main window.
In the right hand side window, we have a different behavior for modules than for classes. In modules, all the elements always show, even when empty, but in classes, only the elements with items in the list will show. I see the difference in code between module_content_list and class_content_list but I cannot fix it.