Update toit.js to latest version.
This is not a great solution, so if someone wants
to pull the latest version from the node packages
repo instead, that would be great.  But this makes
the libs.toit.io so much better that we should use
it as a stop-gap if nobody has time to do it
properly.

There's a slight change to the implementation of
tryImpliedTopLevelFunction, where no indentation
is needed to trigger the "implicit main function"
mode.